pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "cloud-connector-routing-adapter"
version = "0.1.0"
description = "External Cloud Connector routing adapter scaffold currently compatible with VyOS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "requests>=2.31",
]

[project.scripts]
cloud-connector-routing-adapter = "hbr_vyos_adapter.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
